# CAB230 Server Side Application/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Movie API - n10869069",
        "description": "This API provides access to publicly available data on a limited number of movies published from the year 1990 from the Internet Movie Database and other sources. The API endpoints and their usage are described in detail below.",
        "version": "1.0.0",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "n10869069@qut.edu.au"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter JWT Bearer token **_only_**"
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
    "paths": {
        "/movies/search": {
            "get": {
              "tags": [
                "Movies"
              ],
              "summary": "Search movies",
              "description": "Retrieves movies based on search inputs",
              "operationId": "searchMovies",
              "parameters": [
                {
                  "name": "year",
                  "in": "query",
                  "description": "The year to search movies for",
                  "required": false,
                  "schema": {
                    "type": "string"
                  }
                },
                {
                  "name": "page",
                  "in": "query",
                  "description": "The page to search movies for",
                  "required": false,
                  "schema": {
                    "type": "string"
                  }
                },
                {
                  "name": "title",
                  "in": "query",
                  "description": "The title to search movies for",
                  "required": false,
                  "schema": {
                    "type": "string"
                  }
                }
              ],
              "produces": [
                "application/json"
              ],
              "responses": {
                "200": {
                  "description": "Successful operation",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "Error": {
                            "type": "boolean",
                            "enum": [
                              false
                            ]
                          },
                          "Message": {
                            "type": "string",
                            "enum": [
                              "Success"
                            ]
                          },
                          "data": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "title": {
                                  "type": "string",
                                  "example": "Star Trek: First Contact"
                                },
                                "year": {
                                  "type": "integer",
                                  "example": 1996
                                },
                                "imdbID": {
                                  "type": "string",
                                  "example": "tt0117731"
                                },
                                "imdbRating": {
                                  "type": "number",
                                  "example": 7.6
                                },
                                "rottenTomatoesRating": {
                                  "type": "number",
                                  "example": 92
                                },
                                "metacriticRating": {
                                  "type": "number",
                                  "example": 71
                                },
                                "classification": {
                                  "type": "string",
                                  "example": "PG-13"
                                }
                              }
                            }
                          },
                          "pagination": {
                            "type": "object",
                            "properties": {
                              "total": {
                                "type": "integer",
                                "example": 71
                              },
                              "lastPage": {
                                "type": "integer",
                                "example": 7
                              },
                              "perPage": {
                                "type": "integer",
                                "example": 1
                              },
                              "currentPage": {
                                "type": "integer",
                                "example": 2
                              },
                              "from": {
                                "type": "integer",
                                "example": 1
                              },
                              "to": {
                                "type": "integer",
                                "example": 7
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "400": {
                  "description": "Invalid year query parameter. Click on 'Schema' below to see the possible error responses.",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "Error": {
                            "type": "boolean",
                            "enum": [
                              true
                            ]
                          },
                          "Message": {
                            "type": "string",
                            "example": "Invalid year format. Format must be yyyy."
                          }
                        }
                      }
                    }
                  }
                },
                "500": {
                  "description": "Internal server error",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "Error": {
                            "type": "boolean",
                            "enum": [
                              true
                            ]
                          },
                          "Message": {
                            "type": "string",
                            "example": "Error in MySQL query"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
...
        "/people/{id}": {
            "get": {
              "tags": [
                "People"
              ],
              "summary": "Get Person Details",
              "description": "Retrieves details of a person based on their ID.",
              "operationId": "getPersonDetails",
              "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "description": "The person's IMDB ID",
                  "required": true,
                  "schema": {
                    "type": "string"
                  }
                }
              ],
              "security": [
                {
                  "bearerAuth": []
                }
              ],
              "responses": {
                "200": {
                  "description": "Success",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "birthYear": {
                            "type": "integer",
                            "nullable": true
                          },
                          "deathYear": {
                            "type": "integer",
                            "nullable": true
                          },
                          "roles": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "movieName": {
                                  "type": "string"
                                },
                                "movieId": {
                                  "type": "string"
                                },
                                "category": {
                                  "type": "string"
                                },
                                "characters": {
                                  "type": "array",
                                  "items": {
                                    "type": "string"
                                  }
                                },
                                "imdbRating": {
                                  "type": "number"
                                }
                              }
                            }
                          }
                        }
                      },
                      "example": {
                        "name": "Patrick Stewart",
                        "birthYear": 1940,
                        "deathYear": null,
                        "roles": [
                          {
                            "movieName": "Star Trek: First Contact",
                            "movieId": "tt0117731",
                            "category": "Actor",
                            "characters": [
                              "Picard"
                            ],
                            "imdbRating": 7.6
                          }
                        ]
                      }
                    }
                  }
                },
                "400": {
                  "description": "Invalid query parameters. Click on 'Schema' to see possible error responses.",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "error": {
                            "type": "boolean",
                            "enum": [
                              true
                            ]
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      },
                      "example": {
                        "error": true,
                        "message": "Invalid query parameters: year. Query parameters are not permitted."
                      }
                    }
                  }
                },
                "401": {
                  "description": "Unauthorized. Click on 'Examples' below to see the possible error responses.",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "error": {
                            "type": "boolean",
                            "enum": [
                              true
                            ]
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      },
                      "examples": {
                        "missingAuthHeader": {
                          "value": {
                            "error": true,
                            "message": "Missing Authorization header"
                          }
                        },
                        "invalidToken": {
                          "value": {
                            "error": true,
                            "message": "Invalid JWT token"
                          }
                        },
                        "expiredToken": {
                          "value": {
                            "error": true,
                            "message": "Expired JWT token"
                          }
                        }
                      }
                    }
                  }
                },
                "404": {
                  "description": "Not Found",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "error": {
                            "type": "boolean",
                            "enum": [
                              true
                            ]
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      },
                      "example": {
                        "error": true,
                        "message": "Person not found"
                      }
                    }
                  }
                },
                "500": {
                  "description": "Internal server error",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "error": {
                            "type": "boolean",
                            "enum": [
                              true
                            ]
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
...
        "/movies/data/{imdbID}": {
            "get": {
              "tags": [
                "Movies"
              ],
              "summary": "Get movie details by IMDb ID",
              "description": "Retrieves movie details based on the provided IMDb ID.",
              "operationId": "getMovieDetails",
              "parameters": [
                {
                  "name": "imdbID",
                  "in": "path",
                  "description": "The IMDb ID of the movie",
                  "required": true,
                  "schema": {
                    "type": "string"
                  }
                }
              ],
              "responses": {
                "200": {
                  "description": "Successful operation",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "title": {
                            "type": "string",
                            "example": "Star Trek: First Contact"
                          },
                          "year": {
                            "type": "integer",
                            "example": 1996
                          },
                          "runtime": {
                            "type": "integer",
                            "example": 111
                          },
                          "genres": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "country": {
                            "type": "string",
                            "example": "United States"
                          },
                          "principals": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string",
                                  "example": "nm0005772"
                                },
                                "category": {
                                  "type": "string",
                                  "example": "cinematographer"
                                },
                                "name": {
                                  "type": "string",
                                  "example": "Matthew F. Leonetti"
                                },
                                "characters": {
                                  "type": "array",
                                  "items": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          },
                          "ratings": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "source": {
                                  "type": "string",
                                  "example": "Internet Movie Database"
                                },
                                "value": {
                                  "type": "number",
                                  "example": 7.6
                                }
                              }
                            }
                          },
                          "boxoffice": {
                            "type": "string",
                            "example": "92027888"
                          },
                          "poster": {
                            "type": "string",
                            "example": "https://m.media-amazon.com/images/M/MV5BYzMzZmE3MTItODYzYy00YWI5LWFkNWMtZTY5NmU2MDkxYWI1XkEyXkFqcGdeQXVyMjUzOTY1NTc@._V1_SX300.jpg"
                          },
                          "plot": {
                            "type": "string",
                            "example": "The Borg travel back in time intent on preventing Earth's first contact with an alien species. Captain Picard and his crew pursue them to ensure that Zefram Cochrane makes his maiden flight reaching warp speed."
                          }
                        }
                      }
                    }
                  }
                },
                "400": {
                  "description": "Invalid query parameters. Click on 'Schema' to see possible error responses.",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "error": {
                            "type": "boolean",
                            "enum": [
                              true
                            ]
                          },
                          "message": {
                            "type": "string",
                            "example": "Invalid query parameters: year. Query parameters are not permitted."
                          }
                        }
                      }
                    }
                  }
                },
                "404": {
                  "description": "The requested movie could not be found",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "error": {
                            "type": "boolean",
                            "enum": [
                              true
                            ]
                          },
                          "message": {
                            "type": "string",
                            "example": "No record exists of a movie with this ID"
                          }
                        }
                      }
                    }
                  }
                },
                "500": {
                  "description": "Internal server error",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "error": {
                            "type": "boolean",
                            "enum": [
                              true
                            ]
                          },
                          "message": {
                            "type": "string",
                            "example": "Error in MySQL query"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
...
    }
}